+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - [Bugsnag](https://bugsnag.com) compatible logger
+- Constructor for standard logger for errors (`NewStandardErrorLogger`)
 
 
 ## [0.7.1] - 2018-12-22